Address npm cache clear error
With npm 5.x and nodejs 8.x, you might see the following error:

| DEBUG: Executing shell function do_compile
| npm ERR! As of npm@5, the npm cache self-heals from corruption issues
and data extracted from the cache is guaranteed to be valid. If you want
to make sure everything is consistent, use 'npm cache verify' instead.
| npm ERR!
| npm ERR! If you're sure you want to delete the entire cache, rerun
this command with --force.

This might be a npm issue, which has been reported here:
https://github.com/npm/npm/issues/18894

So we add --force to workaround the error until upstream npm fixes it.

Signed-off-by: Yunguo Wei <yunguo.wei@windriver.com>